backend/requirements.txt
flask==3.1.2
flask-cors==5.0.1
bcrypt==5.0.0
gunicorn==25.1.0
python-dotenv==1.2.2
Werkzeug==3.1.5
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
blinker==1.9.0
click==8.3.1
